+++ src/components/AttractionEditor/service.ts
@@ -1,4 +1,4 @@
-import { createLanguagesObject } from "@/src/utils/content";
+import { createLanguagesObject } from "@services/content";
 import { AdminAttraction } from "@api/client/models/AdminAttraction";
 import { CreateAttractionRequest } from "@api/client/models/CreateAttractionRequest";
