+++ src/components/LocationDetailsPage/index.tsx
@@ -2,7 +2,7 @@ import { Location } from "@api/client/models/Location";
 import ROUTES from "@common/routes";
 import Page from "@components/Page";
 import PageTitle from "@components/PageTitle";
-import { getLocalizedLabel } from "@utils/content";
+import { getLocalizedLabel } from "@services/content";
 import { useRouter } from "next/router";
 import { useCallback } from "react";
 import { useTranslations } from "use-intl";
