+++ src/pages/admin/locations/index.tsx
@@ -1,6 +1,6 @@
 import LocationsPage from "@components/LocationsPage";
-import { loadMessages } from "@utils/i18n";
-import withAuth from "@utils/withAuth";
+import { loadMessages } from "@services/i18n";
+import withAuth from "@services/withAuth";
 import { GetStaticProps } from "next";
 
 export const getStaticProps: GetStaticProps = async (context) => ({
