+++ src/pages/registration/index.tsx
@@ -5,8 +5,8 @@ import Button from "@components/Button";
 import FormWrapper from "@components/FormWrapper";
 import { Input } from "@components/InputField";
 import Page from "@components/Page";
-import { loadMessages } from "@utils/i18n";
-import { validateEmail, validatePassword, validateRepeatPassword } from "@utils/validation";
+import { loadMessages } from "@services/i18n";
+import { validateEmail, validatePassword, validateRepeatPassword } from "@services/validation";
 import { GetStaticProps } from "next";
 import { useRouter } from "next/router";
 import React, { FC, FormEvent, useEffect } from "react";
