+++ tsconfig.json
@@ -23,7 +23,7 @@
 			"@components/*": ["./src/components/*"],
 			"@contexts/*": ["./src/contexts/*"],
 			"@hooks/*": ["./src/hooks/*"],
-			"@utils/*": ["./src/utils/*"]
+			"@services/*": ["./src/services/*"]
 		}
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
